refactor and simplify NavLinks
