+++ src/app/Navigation/NavLinks/NavLinks.scss
@@ -1,8 +1,8 @@
-// NavigationLinks.scss
+// NavLinks.scss
 
-@import "../global/styles/config";
+@import "../../../global/styles/config.scss";
 
-.NavigationLinks {
+.NavLinks {
   display: block;
   position: absolute;
   width: 100%;
@@ -18,9 +18,20 @@
   }
 }
 
+.NavLink {
+  color: $black;
+  text-decoration: none;
+  width: 100%;
+  display: block;
+  padding: 12px 0;
+  &-active {
+    background-color: red;
+  }
+}
+
 /* Medium devices (768px and up) */
 @media only screen and (min-width: $bp-m) {
-  .NavigationLinks {
+  .NavLinks {
     display: inline-flex;
     position: relative;
     width: auto;
@@ -30,5 +41,9 @@
     background-color: transparent;
     transform: none;
     transition: initial;
+  }
+
+  .NavLink {
+    padding: 12px 16px;
   }
 }
