--- src/app/Navigation/NavigationLinks/NavigationLink/NavigationLink.js
@@ -1,25 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { NavLink } from "react-router-dom";
-
-import { navigationLink } from "./NavigationLink.scss";
-
-const NavigationLink = ({ path, content = "", active = true }) => (
-  <NavLink to={path} className={navigationLink} active={active}>
-    {content}
-  </NavLink>
-);
-
-export default NavigationLink;
-
-NavigationLink.propTypes = {
-  content: PropTypes.string,
-  path: PropTypes.string,
-  active: PropTypes.bool
-};
-
-NavigationLink.defaultProps = {
-  content: "",
-  path: null,
-  active: true
-};
