--- src/app/Navigation/NavigationLinks/NavigationLink/NavigationLink.scss
@@ -1,18 +1,0 @@
-// NavigationLink.scss
-
-@import "../../../global/styles/config.scss";
-
-.navigationLink {
-  color: $black;
-  text-decoration: none;
-  width: 100%;
-  display: block;
-  padding: 12px 0;
-}
-
-/* Medium devices (768px and up) */
-@media only screen and (min-width: $bp-m) {
-  .navigationLink {
-    padding: 12px 16px;
-  }
-}
